move ut to sql/core
