+++ sql/core/src/test/scala/org/apache/spark/sql/SQLQuerySuite.scala
@@ -22,6 +22,8 @@ import java.net.{MalformedURLException, URL}
 import java.sql.{Date, Timestamp}
 import java.util.concurrent.atomic.AtomicBoolean
 
+import com.google.common.io.Files
+
 import org.apache.spark.{AccumulatorSuite, SparkException}
 import org.apache.spark.scheduler.{SparkListener, SparkListenerJobStart}
 import org.apache.spark.sql.catalyst.util.StringUtils
@@ -33,6 +35,7 @@ import org.apache.spark.sql.execution.datasources.v2.parquet.ParquetScan
 import org.apache.spark.sql.execution.joins.{BroadcastHashJoinExec, CartesianProductExec, SortMergeJoinExec}
 import org.apache.spark.sql.functions._
 import org.apache.spark.sql.internal.SQLConf
+import org.apache.spark.sql.internal.SQLConf.{PARTITION_OVERWRITE_MODE, PartitionOverwriteMode}
 import org.apache.spark.sql.test.{SharedSparkSession, TestSQLContext}
 import org.apache.spark.sql.test.SQLTestData._
 import org.apache.spark.sql.types._
@@ -3190,6 +3193,34 @@ class SQLQuerySuite extends QueryTest with SharedSparkSession {
       checkAnswer(df2, Array(Row(100)))
       val df3 = sql("select case when 1=2 then 1 else 1.000000000000000000000001 end / 10")
       checkAnswer(df3, Array(Row(new java.math.BigDecimal("0.100000000000000000000000100"))))
+    }
+  }
+
+  test("SPARK-29037: Spark gives duplicate result when an application was killed") {
+    withSQLConf(PARTITION_OVERWRITE_MODE.key -> PartitionOverwriteMode.DYNAMIC.toString) {
+      withTable("test") {
+        sql("create table test(id int, p1 int, p2 int) using parquet partitioned by (p1, p2)")
+        sql("insert overwrite table test partition(p1=1,p2) select 1,3")
+        val df = sql("select id from test where p1=1 and p2=3")
+        checkAnswer(df, Array(Row(1)))
+
+        val warehouse = SQLConf.get.warehousePath.split(":").last
+        val tblPath = Array(warehouse, "test").mkString(File.separator)
+        val taskAttemptPath = Array(tblPath, "_temporary", "0", "task_20190914232019_0000_m_000000",
+          "p1=1", "p2=3").mkString(File.separator)
+        new File(taskAttemptPath).mkdirs()
+
+        val tblResult = new File(Array(tblPath, "p1=1", "p2=3").mkString(File.separator))
+        val tFile = tblResult.list((_: File, name: String) => !name.startsWith(".")).apply(0)
+        val from = new File(tblResult.getAbsolutePath + File.separator + tFile)
+        val to = new File(taskAttemptPath + File.separator + tFile)
+        Files.copy(from, to)
+
+        sql("insert overwrite table test partition(p1=1,p2) select 2, 3")
+        assert(tblResult.list((_: File, name: String) => !name.startsWith(".")).size === 1)
+        val df2 = sql("select id from test where p1=1 and p2=3")
+        checkAnswer(df2, Array(Row(2)))
+      }
     }
   }
 }
